MDEV-34481 optimize away waiting for owned by prepared xa non-unique index record
Two transaction that update a non-unique-only index table could not be
isolated on slave when on slave they used different non-unique indexes
than on master.
A hang could be seen in case the first of the two is a prepared XA

--connection slave_worker_1

  xa start 'xid'; /* ... lock here ... */ ; xa prepare 'xid'

while the 2nd being of any kind including of normal type

--connection slave_worker_2

  begin; /* ... get lock ... => wait/hang...error out  */

was unable to wait up for the conflicting lock, even though
the XA transaction did not really lock target records of the 2nd.

This type of hang was caused by a chosen method the 2nd transaction
employs to reach the target record, which is the index scan. The scanning
orthodoxically just could not step over a record in the way that was
locked by the XA.
However as the in-the-way record can not be targeted by the 2nd
transaction, otherwise the transactions would have sensed the conflict
back on master, *and* the other possibility of collecting extra locks by the
'xid' on non-modified records is tacked by MDEV-33454/MDEV-34466,
the non-unique index scanning must not panic at seeing a timeout error
from the engine. Instead the scanning would
just proceed to next possibly free index records of the same key value and
ultimately must reach the target one.
Also on the way to its target all busy records belonging to
earlier (binlog order) prepared XA transactions need not be tried locking
by the current non-unique index scanning transaction.

This patch implements the plan for Innodb.
The server layer expects the engine to return a new handler error when
an attempt to wait for a conflicting lock that belongs to a
transaction in prepared state is made. The server proceeds to any next
available record of a chosen non-unique key.

Note this work partly implements MDEV-33455 agenda for the ROW binlog
format of replicated DML on non-unique-index-only table.