MDEV-34481 Revised HA_ERR_LOCK_WAIT_TIMEOUT method to introduce a non-rolling-back
Earlier introduced RGI::exec_flags is reused to inform engine
on that the current replication row-based event requires a READ-PAST
report back in case the event is on a non-unique index and hits
a lock in the way from a prepared user XA.
When that's the case Innodb internally generates DB_LOCK_READ_PAST
"error" state which translates to a new HA_ERR_NUK_LOCKED code
returned to the slave applier's server layer.
The new Innodb status is handled largely as DB_SUCCESS except
no actual locking is done.